--- a/gomponents_test.go
+++ b/gomponents_test.go
@@ -54,6 +54,16 @@ func TestAttr(t *testing.T) {
 			t.FailNow()
 		}
 	})
+
+	t.Run("escapes attribute names", func(t *testing.T) {
+		a := g.Attr(`"><script`)
+		assert.Equal(t, ` &#34;&gt;&lt;script`, a)
+	})
+
+	t.Run("escapes attribute names and values", func(t *testing.T) {
+		a := g.Attr(`"><script`, `hat"><script`)
+		assert.Equal(t, ` &#34;&gt;&lt;script="hat&#34;&gt;&lt;script"`, a)
+	})
 }
 
 func BenchmarkAttr(b *testing.B) {
@@ -129,6 +139,11 @@ func TestEl(t *testing.T) {
 		e := g.El("div")
 		err := e.Render(&erroringWriter{})
 		assert.Error(t, err)
+	})
+
+	t.Run("escapes tag names", func(t *testing.T) {
+		e := g.El(`div><script`)
+		assert.Equal(t, `<div&gt;&lt;script></div&gt;&lt;script>`, e)
 	})
 }