Because this can be a place of injection if untrusted data is passed, escape all attribute values.

This slows down rendering a bit, but we're still talking tiny increments compared to everything else.

### Before

```
$ make benchmark
go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/maragudk/gomponents
BenchmarkAttr/boolean_attributes-8         	13095948	        91.92 ns/op
BenchmarkAttr/name-value_attributes-8      	 9124303	       131.5 ns/op
BenchmarkEl/normal_elements-8              	 6384505	       187.4 ns/op
PASS
ok  	github.com/maragudk/gomponents	5.124s
```

### After

```
$ make benchmark
go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/maragudk/gomponents
BenchmarkAttr/boolean_attributes-8         	11524586	        92.07 ns/op
BenchmarkAttr/name-value_attributes-8      	 7606478	       157.0 ns/op
BenchmarkEl/normal_elements-8              	 6390315	       187.0 ns/op
PASS
ok  	github.com/maragudk/gomponents	4.000s
```

See #74.