@oderwat 2021-05-16
I think this is the right way to go. XSS is one thing, strange bugs are the other. Automagic escaping really helps to not break stuff by accident.