@markuswustenberg 2021-05-17
@oderwat thanks for the feedback. 😊 Have you seen the discussion in #74 as well? I think some of this will still be changed, I'm not sure I'm keeping attribute name escaping.