@oderwat 2021-05-17
Yes, I saw it. I actually think that the system should probably "panic" for invalid attribute names. Alternative mangle them into something valid (replacing invalid characters with an _ maybe). Valid could be everything except "whitespace" and `/>"'=`.