@markuswustenberg 2021-05-18
> Yes, I saw it. I actually think that the system should probably "panic" for invalid attribute names. Alternative mangle them into something valid (replacing invalid characters with an _ maybe). Valid could be everything except "whitespace" and `/>"'=`.

Hmm. I think I'd rather not panic, because I would be very sad as a library user if the library panicked and I didn't expect it. I think I'll hope that the developer knows what s/he is doing instead, and let them misuse the library. I think that's the least surprising thing to do.

If this turns out to be an issue, I can revisit the decision.